[update] 型の修正

+++ src/PointCalculate.cpp
@@ -10,15 +10,15 @@ PointCalculate::PointCalculate(vector<vector<Tile>>& tile,vector<vector<Agent>>&
 
 Pos PointCalculate::makePos(int x,int y){
     Pos p;
-    p.x=x;
-    p.y=y;
+    p.x=static_cast<unsigned>(x);
+    p.y=static_cast<unsigned>(y);
     return p;
 }
 
 int PointCalculate::getTilePoints(int color){
     int tilePoint=0;
-    for(int x=0;x<width;x++){
-        for(int y=0;y<height;y++){
+    for(unsigned int x=0;x<width;x++){
+        for(unsigned int y=0;y<height;y++){
             if(tile[y][x].color==color){
                 tilePoint+=tile[y][x].point;//現在見るべき色のタイルのみを計算に入れる
 
@@ -32,10 +32,10 @@ int PointCalculate::getAreaPoints(int color){
     //領域データの設定
     areaData.clear();
     checkedData.clear();
-    for(int y=0;y<height;y++){//配列用意の順番的にxからループを始める
+    for(unsigned int y=0;y<height;y++){//配列用意の順番的にxからループを始める
         vector<bool> tmpDataB;
         vector<int> tmpDataI;
-        for(int x=0;x<width;x++){
+        for(unsigned int x=0;x<width;x++){
             tmpDataB.push_back(false);
             tmpDataI.push_back(0);
         }
@@ -43,19 +43,19 @@ int PointCalculate::getAreaPoints(int color){
         checkedData.push_back(tmpDataI);
     }
 
-    for(int y=1;y<height-1;y++){//フィールドの端のマスは領域ポイントを取らないためループから省く
-        for(int x=1;x<width-1;x++){//横に見ていく、全部見たら次の列、という形
+    for(unsigned int y=1;y<height-1;y++){//フィールドの端のマスは領域ポイントを取らないためループから省く
+        for(unsigned int x=1;x<width-1;x++){//横に見ていく、全部見たら次の列、という形
             if(tile[y][x].color!=color&&checkedData[y][x]!=-2){//自タイルが置かれていない&&非領域エリアでないなら検証を行う
-                if(checkArea(makePos(x,y),color)){//検証結果が１...-1とチェックしたマスはareaDataへと加える
-                    for(int y=0;y<height;y++){
-                        for(int x=0;x<width;x++){
+                if(checkArea(makePos(static_cast<int>(x),static_cast<int>(y)),color)){//検証結果が１...-1とチェックしたマスはareaDataへと加える
+                    for(unsigned int y=0;y<height;y++){
+                        for(unsigned int x=0;x<width;x++){
                             if(checkedData[y][x]==-1)areaData[y][x]=1;
                         }
                     }
                 }
                 else {//検証結果が０...-1とチェックしたマスは非領域エリア(-2)としてcheckedDataを更新、処理を減らすため
-                    for(int y=0;y<height;y++){
-                        for(int x=0;x<width;x++){
+                    for(unsigned int y=0;y<height;y++){
+                        for(unsigned int x=0;x<width;x++){
                             if(checkedData[y][x]==-1)checkedData[y][x]=-2;
                         }
                     }
@@ -66,8 +66,8 @@ int PointCalculate::getAreaPoints(int color){
 
     //完成したareaDataを用いて領域ポイントを計算する
     int areaPoint=0;
-    for(int x=0;x<width;x++){
-        for(int y=0;y<height;y++){
+    for(unsigned int x=0;x<width;x++){
+        for(unsigned int y=0;y<height;y++){
             if(areaData[y][x]==1)areaPoint+=abs(tile[y][x].point);//絶対値で計算
         }
     }
@@ -102,7 +102,7 @@ int PointCalculate::checkArea(Pos checkPos,int color){
         else if(checkedData[checkPos.y+looking[i].y][checkPos.x+looking[i].x]!=-1){//元のマスを見ないようにする
             //見ているマスが端に接しておらず、自タイルも置かれていない場合...領域の中身になる可能性があるため、チェックして再帰、そのマスからこの関数を呼び出す
             checkedData[checkPos.y][checkPos.x]=-1;//領域の中は-1でチェック
-            int result=checkArea(makePos(checkPos.x+looking[i].x,checkPos.y+looking[i].y),color);
+            int result=checkArea(makePos(static_cast<int>(checkPos.x+looking[i].x),static_cast<int>(checkPos.y+looking[i].y)),color);
             //再帰中にfieldの端に接した場合...result==0,領域ポイントを取らないため、ループを抜けて次の検証に移る
             if(result==0)return result;
         }
